+++ src/index.html
@@ -1,13 +1,12 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Bear Watcher</title>
-</head>
-<body>
-  <div id="app"></div>
-  <button type="button" id="all-bears" class="btn btn-success">See All The Bears</button>
-  <div id="app2"></div>
-</body>
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Bear Watcher</title>
+  </head>
+  <body>
+    <div id="app"></div>
+    <div id="app2" class="d-flex flex-wrap"></div>
+  </body>
 </html>
