+++ src/javascripts/data/bears.js
@@ -1,0 +1,16 @@
+const bears = [
+  {
+    name: 'Revel',
+    picture: 'https://www.gannett-cdn.com/presto/2018/08/22/USAT/76151925-b45a-45b6-b62d-9fe0b3cc4c77-XXX_Trevor_Hughes_Yellowstone-Teton_August2018_2200.JPG?crop=5199,2960,x0,y1043&width=3200&height=1822&format=pjpg&auto=webp',
+  },
+  {
+    name: 'Bubba',
+    picture: 'https://static01.nyt.com/images/2020/09/30/multimedia/00xp-fat-bear-week3/00xp-fat-bear-week3-mobileMasterAt3x.jpg',
+  },
+  {
+    name: 'Popscicle',
+    picture: 'https://www.outsideonline.com/sites/default/files/styles/full-page/public/2018/10/02/otis-mid-chew_h.jpg?itok=UScDb-fS',
+  },
+];
+
+export default bears;
