+++ src/styles/main.scss
@@ -12,5 +12,9 @@ body {
 }
 
  #card1 {
-   object-fit: scale-down;
+   margin: 15px;
+}
+
+.card-img-top {
+  border-radius: 40%;
 }
